specify uint8array payload type

+++ solver/src/vaaAuctionRelayer/app.ts
@@ -81,9 +81,7 @@ async function main(argv: string[]) {
 
             // Start a new auction if this is a fast VAA.
             logicLogger.debug(`Attempting to parse FastMarketOrder, sequence=${parsed.sequence}`);
-            const fastOrder = utils.tryParseFastMarketOrder(
-                Buffer.from(parsed.payload as Uint8Array),
-            );
+            const fastOrder = utils.tryParseFastMarketOrder(Buffer.from(parsed.payload));
             if (fastOrder !== undefined) {
                 const unprocessedTxns = await utils.handlePlaceInitialOffer(
                     connection,
@@ -102,9 +100,7 @@ async function main(argv: string[]) {
             }
         } else {
             logicLogger.debug(`Attempting to parse SlowOrderResponse, sequence=${parsed.sequence}`);
-            const slowOrderResponse = utils.tryParseSlowOrderResponse(
-                Buffer.from(parsed.payload as Uint8Array),
-            );
+            const slowOrderResponse = utils.tryParseSlowOrderResponse(Buffer.from(parsed.payload));
             if (slowOrderResponse !== undefined) {
                 const unprocessedTxns = await utils.handleSettleAuction(
                     connection,
